backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-groq
pydantic
websockets
python-dotenv
